+++ examples/demo_kraken_futures.py
@@ -7,7 +7,7 @@ associated with this software.
 from cryptofeed.callback import TickerCallback, TradeCallback, BookCallback
 from cryptofeed import FeedHandler
 from cryptofeed.exchanges import KrakenFutures
-from cryptofeed.defines import L2_BOOK, BID, ASK, TRADES, TICKER, FUNDING
+from cryptofeed.defines import L2_BOOK, BID, ASK, TRADES, TICKER, FUNDING, OPEN_INTEREST
 
 
 async def trade(feed, pair, order_id, timestamp, side, amount, price):
@@ -26,11 +26,15 @@ async def funding(**kwargs):
     print(f"Funding: {kwargs}")
 
 
+async def oi(feed, pair, open_interest, timestamp):
+    print(f'Timestamp: {timestamp} Feed: {feed} Pair: {pair} open interest: {open_interest}')
+
+
 def main():
     fh = FeedHandler()
 
-    config = {TRADES: ['PI_XBTUSD'], TICKER: ['PI_XBTUSD', 'PI_ETHUSD'], L2_BOOK: ['PI_XBTUSD'], FUNDING: ['PI_XBTUSD']}
-    fh.add_feed(KrakenFutures(config=config, callbacks={FUNDING: funding, TICKER: TickerCallback(ticker), TRADES: TradeCallback(trade), L2_BOOK: BookCallback(book)}))
+    config = {OPEN_INTEREST: ['PI_XBTUSD', 'PI_ETHUSD'], TRADES: ['PI_XBTUSD'], TICKER: ['PI_XBTUSD', 'PI_ETHUSD'], L2_BOOK: ['PI_XBTUSD'], FUNDING: ['PI_XBTUSD']}
+    fh.add_feed(KrakenFutures(config=config, callbacks={OPEN_INTEREST: oi, FUNDING: funding, TICKER: TickerCallback(ticker), TRADES: TradeCallback(trade), L2_BOOK: BookCallback(book)}))
 
     fh.run()