+++ CHANGES.md
@@ -8,6 +8,7 @@
   * Change: Instrument removed in favor of open interest
   * Bugfix: Huobi feedhandlers not properly setting forced indicator for book updates, breaking deltas
   * Bugfix: Some kraken futures funding fields not always populated
+  * Feature: Open interest updates for kraken futures
 
 ### 1.2.0 (2020-01-18)
   * Feature: New exchange: Binance Futures